ingestion/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ingestion"
version = "1.0.0"
description = "Domain ingestion pipeline — CZDS zone files + OpenINTEL snapshots → R2 Parquet + PostgreSQL"
requires-python = ">=3.12"
dependencies = [
    "polars>=0.20",
    "boto3>=1.34",
    "requests>=2.31",
    "pydantic-settings>=2.2",
    "psycopg2-binary>=2.9",
    "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["."]
include = ["ingestion*"]

[tool.setuptools.package-data]
"ingestion.databricks.notebooks" = ["*.py"]

[project.scripts]
ingestion = "ingestion.cli:main"
